+++ Dockerfile-service
@@ -1,0 +1,4 @@
+FROM python:3-onbuild
+RUN apt-get update && apt-get install -y wget dnsutils
+ENV PYTHONUNBUFFERED doubleplusgood
+CMD ["nameko", "run", "src.service", "--config", "config.yaml"]
