+++ requirements.txt
@@ -1,30 +1,42 @@
+Flask==0.11.1
+Jinja2==2.8
+Mako==1.0.4
+MarkupSafe==0.23
+PyYAML==3.11
+SQLAlchemy==1.1.0b2
+SQLAlchemy-Utils==0.32.9
+Werkzeug==0.11.5
+alembic==0.8.6
 amqp==1.4.9
+aniso8601==1.1.0
 anyjson==0.3.3
 appnope==0.1.0
+backports.shutil-get-terminal-size==1.0.0
+click==6.6
 decorator==4.0.9
 eventlet==0.18.4
 gnureadline==6.3.3
 greenlet==0.4.9
-ipdb==0.9.3
-ipython==4.1.2
+ipdb==0.10.0
+ipdbplugin==1.4.5
+ipython==4.2.0
 ipython-genutils==0.1.0
+itsdangerous==0.24
 kombu==3.0.35
 mock==2.0.0
 nameko==2.3.0
-path.py==8.2.1
-pbr==1.9.1
+nose==1.3.7
+path.py==8.2
+pbr==1.9.0
 pexpect==4.0.1
 pickleshare==0.7.2
 ptyprocess==0.5.1
 py==1.4.31
-pytest==2.9.1
-python-http-client==1.2.3
-PyYAML==3.11
+python-dateutil==2.5.3
+python-editor==1.0
+pytz==2016.6.1
 requests==2.9.1
 sendgrid==2.2.1
 simplegeneric==0.8.1
 six==1.10.0
-smtpapi==0.3.1
 traitlets==4.2.1
-Werkzeug==0.11.8
-wheel==0.24.0
